requirements.txt
cloudpathlib==0.10.0
Cython==0.29.28
ffmpeg-python==0.2.0
imageio==2.16.1
imageio-ffmpeg==0.4.5
librosa==0.9.1
madmom==0.16.1
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.5.64
pre-commit==2.18.1
pydub==0.25.1
torch==1.10.2
tqdm==4.63.1